New waiver model, remove old waiver model
