+++ talentmap_api/language/models.py
@@ -97,34 +97,3 @@ class Qualification(StaticRepresentationModel):
         managed = True
         ordering = ["language__code"]
         unique_together = (('language', 'reading_proficiency', 'spoken_proficiency'))
-
-
-class Waiver(StaticRepresentationModel):
-    '''
-    The waiver model represents a language requirement waiver for a particular position
-    assignment, along with the status
-    '''
-
-    class Type(DjangoChoices):
-        partial = ChoiceItem("partial")
-        full = ChoiceItem("full")
-
-    class Status(DjangoChoices):
-        approved = ChoiceItem("approved")
-        requested = ChoiceItem("requested")
-        denied = ChoiceItem("denied")
-
-    type = models.TextField(default=Type.full, choices=Type.choices)
-    status = models.TextField(default=Status.requested, choices=Status.choices)
-
-    user = models.ForeignKey("user_profile.UserProfile", on_delete=models.CASCADE, related_name="language_waivers")
-    position = models.ForeignKey("position.Position", related_name="language_waivers")
-    bidcycle = models.ForeignKey("bidding.BidCycle", related_name="language_waivers")
-    language = models.ForeignKey("language.Language", related_name="waivers")
-
-    request_date = models.DateField(auto_now_add=True)
-    decision_date = models.DateField(null=True)
-
-    class Meta:
-        managed = True
-        ordering = ["request_date"]
